third_party/blink/web_tests/external/wpt/fullscreen/api/historical-expected.txt

This is a testharness.js-based test.
Found 17 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] <video> member must not be supported: onwebkitfullscreenchange
  assert_false: expected false got true
[FAIL] Document member must not be supported: onwebkitfullscreenchange
  assert_false: expected false got true
[FAIL] <video> member must not be supported: onwebkitfullscreenerror
  assert_false: expected false got true
[FAIL] Document member must not be supported: onwebkitfullscreenerror
  assert_false: expected false got true
[FAIL] Document member must not be supported: webkitCurrentFullScreenElement (uppercase S)
  assert_false: expected false got true
[FAIL] Document member must not be supported: webkitFullscreenElement
  assert_false: expected false got true
[FAIL] Document member must not be supported: webkitFullscreenEnabled
  assert_false: expected false got true
[FAIL] Document member must not be supported: webkitIsFullScreen (uppercase S)
  assert_false: expected false got true
[FAIL] <video> member must not be supported: webkitRequestFullScreen (uppercase S)
  assert_false: expected false got true
[FAIL] <video> member must not be supported: webkitRequestFullscreen
  assert_false: expected false got true
[FAIL] <video> member must not be supported: webkitDisplayingFullscreen
  assert_false: expected false got true
[FAIL] <video> member must not be supported: webkitEnterFullScreen (uppercase S)
  assert_false: expected false got true
[FAIL] <video> member must not be supported: webkitEnterFullscreen
  assert_false: expected false got true
[FAIL] <video> member must not be supported: webkitExitFullScreen (uppercase S)
  assert_false: expected false got true
[FAIL] <video> member must not be supported: webkitExitFullscreen
  assert_false: expected false got true
[FAIL] Document member must not be supported: webkitExitFullscreen
  assert_false: expected false got true
[FAIL] <video> member must not be supported: webkitSupportsFullscreen
  assert_false: expected false got true
Harness: the test ran to completion.
